Writing, CHomP, script for chomp automation
I’m not sure if I wrote anything since last commit but I added chomp to
the directory so that it will be synced. You have to install it before
running the script.py.

script.py automates analyzing the homology. Give it a folder of png
files (as output by the grey scott simulator). It uses
chomp-greyscale-to-cubical to convert the pngs to a txt that
chomp-cubical can analyze. It writes the results of the homology to a
csv in the format:
name-step, betti-0, betti-1, betti-2